+++ packages/generator-electrode/generators/app/templates/src/client/routes.jsx
@@ -1,7 +1,5 @@
 import React from "react";
-import {Route} from "react-router";
+import { Route } from "react-router";
 import Home from "./components/home";
 
-export const routes = (
-  <Route path="/" component={Home}/>
-);
+export const routes = <Route path="/" component={Home} />;