+++ CHANGELOG.md
@@ -1,10 +1,194 @@
-# 6/21/2017
+# 6/29/2017
 
 ## Packages
 
 ### Directly Updated
 
+-   electrode-redux-router-engine@1.4.5 `(1.4.4 => 1.4.5)`
+-   generator-electrode@3.1.1 `(3.1.0 => 3.1.1)`
+
+### Lerna Updated
+
+-   electrode-archetype-react-app@3.1.3 `(3.1.2 => 3.1.3)`
+-   electrode-archetype-react-app-dev@3.1.3 `(3.1.2 => 3.1.3)`
+-   electrode-webpack-reporter@0.3.10 `(0.3.9 => 0.3.10)`
+
+## Commits
+
+-   `packages/electrode-redux-router-engine`
+
+    -   [patch][bug] Escape **PRELOADED_STATE**. Fix for #441 ([#442](https://github.com/electrode-io/electrode/pull/442)) [commit](http://github.com/electrode-io/electrode/commit/161fb83b605f140d27a537b0e8e3e7564d0fac11)
+
+-   `packages/generator-electrode`
+
+    -   no git init for demo-app ([#445](https://github.com/electrode-io/electrode/pull/445)) [commit](http://github.com/electrode-io/electrode/commit/945b25b8b583c6340e34c2685a829ed10d246492)
+
+-   `docs`
+
+    -   gitbook: [patch] Update docs form training feedback ([#443](https://github.com/electrode-io/electrode/pull/443)) [commit](http://github.com/electrode-io/electrode/commit/cef5d8d8bb197765e497add66feac65d73a93c5c)
+    -   update readme [commit](http://github.com/electrode-io/electrode/commit/1932de00a20c3fac630cc043d0bf30c47c7b4794)
+
+# 6/26/2017
+
+## Packages
+
+-   electrode-archetype-react-app@3.1.2 `(3.1.1 => 3.1.2)`
+-   electrode-archetype-react-app-dev@3.1.2 `(3.1.1 => 3.1.2)`
+-   generator-electrode@3.1.0 `(3.0.4 => 3.1.0)`
+
+## Commits
+
+-   `packages/electrode-archetype-react-app[-dev]`
+
+    -   [patch] fix eslint rules for server tests ([#439](https://github.com/electrode-io/electrode/pull/439)) [commit](http://github.com/electrode-io/electrode/commit/e28b12bdb3f021d2c1569e0ca4ff1e3d097cdf3d)
+    -   [chore] enclose node env in ' [commit](http://github.com/electrode-io/electrode/commit/5c48e4d063f67d1cc0d8cfb8014a018a3bb9b9ab)
+
+-   `packages/generator-electrode`
+
+    -   [minor] improve demos included in the generated app ([#438](https://github.com/electrode-io/electrode/pull/438)) [commit](http://github.com/electrode-io/electrode/commit/8292cb39157a7473323698e55b01a659d68d2189)
+
+# 6/24/2017
+
+## Packages
+
+-   electrode-archetype-react-app@3.1.1 `(3.0.5 => 3.1.1)`
+-   electrode-archetype-react-app-dev@3.1.1 `(3.0.5 => 3.1.1)`
+-   electrode-archetype-react-component@3.0.3 `(3.0.2 => 3.0.3)`
+-   electrode-archetype-react-component-dev@3.0.3 `(3.0.2 => 3.0.3)`
+-   generator-electrode@3.0.4 `(3.0.3 => 3.0.4)`
+
+## Commits
+
+-   `packages/electrode-archetype-react-app[-dev]`
+
+    -   [patch] only optimize CSS for production build ([#436](https://github.com/electrode-io/electrode/pull/436)) [commit](http://github.com/electrode-io/electrode/commit/7ab6f54defae2995becef5b9e5a1d154774a7ab8)
+    -   [minor] update mocha ([#434](https://github.com/electrode-io/electrode/pull/434)) [commit](http://github.com/electrode-io/electrode/commit/286d79aa4642ac57fe8c1346cc48069e42dfba6d)
+    -   [patch] update dep extract-text-plugin ([#433](https://github.com/electrode-io/electrode/pull/433)) [commit](http://github.com/electrode-io/electrode/commit/03c8eda0c78fc87fbba545562adc52ba22ae03ae)
+    -   [patch] fix startAppServer [commit](http://github.com/electrode-io/electrode/commit/cca48935944e809312a3e9572a33ea73dc555970)
+    -   [patch] update dep, xclap, sw-precache ([#431](https://github.com/electrode-io/electrode/pull/431)) [commit](http://github.com/electrode-io/electrode/commit/31a350a5effb8fe9a70f147fa83d476a53df434e)
+    -   [patch] update the deprecated isparta-loader ([#428](https://github.com/electrode-io/electrode/pull/428)) [commit](http://github.com/electrode-io/electrode/commit/013c14b521b720d4211bc2999e3935624a2ec0d5)
+    -   [patch] add xclai-cli to archetype app ([#426](https://github.com/electrode-io/electrode/pull/426)) [commit](http://github.com/electrode-io/electrode/commit/01416b5f5eff2060a99a5d55e9c00e27bcc55484)
+    -   Fix eslint errors [commit](http://github.com/electrode-io/electrode/commit/4f132586b81dab1ed1e7a40f259f77db5a09dfaf)
+    -   Debug server side components with Chrome Developer Tools (DevTools) [commit](http://github.com/electrode-io/electrode/commit/eb8124b0baf8b0af8d86f743e70a365db5f7423f)
+    -   code refactor [commit](http://github.com/electrode-io/electrode/commit/533b94138e3465ad7f8b2ef5509703ff8a5f824d)
+    -   fix typo [commit](http://github.com/electrode-io/electrode/commit/c2086c13bf7e6e628dfae08a8bc29cd0164a463a)
+    -   added missing variable [commit](http://github.com/electrode-io/electrode/commit/b1b0a0fca7c098d7a4f02e274ef183b36be99edd)
+    -   remove extractTextPlugin in hot mode [commit](http://github.com/electrode-io/electrode/commit/19c6e08da5b371a39c6bed3c97407a4c6be5541f)
+    -   [minor] update istanbul ([#435](https://github.com/electrode-io/electrode/pull/435)) [commit](http://github.com/electrode-io/electrode/commit/415ea039cb2f6f309a032b3c838eb384a99919de)
+
+-   `packages/electrode-archetype-react-component[-dev]`
+
+    -   [patch] improve resolving xclap ([#432](https://github.com/electrode-io/electrode/pull/432)) [commit](http://github.com/electrode-io/electrode/commit/1d4051cfd1cae8e29f21ff6795626be96923c3ae)
+
+-   `packages/generator-electrode`
+
+    -   [patch][bug] fix license prompt for app generator, use path. ([#429](https://github.com/electrode-io/electrode/pull/429)) [commit](http://github.com/electrode-io/electrode/commit/0461a59dfd12c6bd2ed9f3a022744908320d1f1e)
+    -   [patch] change autoSSR message ([#425](https://github.com/electrode-io/electrode/pull/425)) [commit](http://github.com/electrode-io/electrode/commit/7212b70180e39421d2a5774e4ef3ff8dd9edab77)
+    -   rename clap.js to xclap.js ([#427](https://github.com/electrode-io/electrode/pull/427)) [commit](http://github.com/electrode-io/electrode/commit/4030b1cc626575c4d2f8736f36804c0c82352478)
+    -   Debug server side components with Chrome Developer Tools (DevTools) [commit](http://github.com/electrode-io/electrode/commit/eb8124b0baf8b0af8d86f743e70a365db5f7423f)
+
+-   `samples/universal-react-node`
+
+    -   [chore] remove xlap-cli from sample [commit](http://github.com/electrode-io/electrode/commit/1efdcf86a9934cea9137eff09d60255b967693cf)
+    -   [chore] default samples to pull from local package ([#430](https://github.com/electrode-io/electrode/pull/430)) [commit](http://github.com/electrode-io/electrode/commit/acf58194afbedc7819ebe32371b9ce03311d8820)
+
+-   `docs`
+
+    -   update readme [commit](http://github.com/electrode-io/electrode/commit/f195c72c174fd13d8fd9800c03d79ba73356bb95)
+
+-   `MISC`
+
+    -   update contributing guide [commit](http://github.com/electrode-io/electrode/commit/84e5eb682f3b2b9b0c73d34f284fc0562dc8a0db)
+    -   update lerna [commit](http://github.com/electrode-io/electrode/commit/c4e70ae0dc8fae527edabc2ea3a8a6b03310ba54)
+
+# 6/22/2017
+
+## Packages
+
+-   electrode-archetype-react-app@3.0.5 `(3.0.3 => 3.0.5)`
+-   electrode-archetype-react-app-dev@3.0.5 `(3.0.3 => 3.0.5)`
+-   electrode-archetype-react-component@3.0.2 `(3.0.1 => 3.0.2)`
+-   electrode-archetype-react-component-dev@3.0.2 `(3.0.1 => 3.0.2)`
+-   electrode-auto-ssr@1.0.3 `(1.0.2 => 1.0.3)`
+-   electrode-react-webapp@1.6.3 `(1.6.2 => 1.6.3)`
+-   electrode-redux-router-engine@1.4.4 `(1.4.3 => 1.4.4)`
+-   electrode-webpack-reporter@0.3.9 `(0.3.8 => 0.3.9)`
+-   generator-electrode@3.0.3 `(3.0.2 => 3.0.3)`
+
+## Commits
+
+-   `packages/electrode-archetype-react-app[-dev]`
+
+    -   [patch] fix devHostname [commit](http://github.com/electrode-io/electrode/commit/bcc2de27ea9382dd2b7256c13732029ae6f08bc2)
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+    -   [patch][chore] clean up env get ([#419](https://github.com/electrode-io/electrode/pull/419)) [commit](http://github.com/electrode-io/electrode/commit/3a91a830516d88e455bb90a97babe04a3464d5b1)
+    -   [chore] add test mocha opts and nyc coverage config [commit](http://github.com/electrode-io/electrode/commit/6aa8bfcacb9f4d89309e321b2f0033ad63deaf7d)
+
+-   `packages/electrode-archetype-react-component[-dev]`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+
+-   `packages/electrode-auto-ssr`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+    -   [chore] add test mocha opts and nyc coverage config [commit](http://github.com/electrode-io/electrode/commit/6aa8bfcacb9f4d89309e321b2f0033ad63deaf7d)
+
+-   `packages/electrode-react-webapp`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+    -   [chore] add test mocha opts and nyc coverage config [commit](http://github.com/electrode-io/electrode/commit/6aa8bfcacb9f4d89309e321b2f0033ad63deaf7d)
+
+-   `packages/electrode-redux-router-engine`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+    -   [chore] add test mocha opts and nyc coverage config [commit](http://github.com/electrode-io/electrode/commit/6aa8bfcacb9f4d89309e321b2f0033ad63deaf7d)
+
+-   `packages/electrode-webpack-reporter`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+
+-   `packages/generator-electrode`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+    -   [patch] check component generators is invoked inside app. ([#422](https://github.com/electrode-io/electrode/pull/422)) [commit](http://github.com/electrode-io/electrode/commit/7fe029720c8fa261ca91da6e1975670300431680)
+    -   [chore] add test mocha opts and nyc coverage config [commit](http://github.com/electrode-io/electrode/commit/6aa8bfcacb9f4d89309e321b2f0033ad63deaf7d)
+
+-   `samples/demo-component`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+
+-   `samples/electrode-demo-index`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+
+-   `samples/universal-material-ui`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+
+-   `samples/universal-react-node`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+
+-   `docs`
+
+    -   [patch] rename clap.js to xclap.js ([#424](https://github.com/electrode-io/electrode/pull/424)) [commit](http://github.com/electrode-io/electrode/commit/2da5a892a7b828f2120b3b23d354fd807c7de2ea)
+
+-   `MISC`
+
+    -   [chore] consolidate 6/21 changelog [commit](http://github.com/electrode-io/electrode/commit/16ec599201c66911273ad2305a6de666fc01b5d9)
+
+# 6/21/2017
+
+Update and fix dependencies after major releases.
+
+## Packages
+
+### Directly Updated
+
+-   electrode-archetype-react-app-dev@3.0.2 `(3.0.1 => 3.0.2)`
+-   electrode-archetype-react-app@3.0.2 `(3.0.1 => 3.0.2)`
 -   electrode-webpack-reporter@0.3.8 `(0.3.7 => 0.3.8)`
+-   generator-electrode@3.0.2 `(3.0.1 => 3.0.2)`
 
 ### Lerna Updated
 
@@ -13,17 +197,13 @@
 
 ## Commits
 
+-   `packages/electrode-archetype-react-app[-dev]`
+
+    -   [patch] update peer dep ([#415](https://github.com/electrode-io/electrode/pull/415)) [commit](http://github.com/electrode-io/electrode/commit/3799989a37482c6b094c17f3e2f1e17e283a2b72)
+
 -   `packages/electrode-webpack-reporter`
 
     -   [patch] fix webpack-reporter ([#417](https://github.com/electrode-io/electrode/pull/417)) [commit](http://github.com/electrode-io/electrode/commit/f57955582d1b4243f75ab8f7ec441013df71aa04)
-
-# 6/21/2017
-
-## Packages
-
--   generator-electrode@3.0.2 `(3.0.1 => 3.0.2)`
-
-## Commits
 
 -   `packages/generator-electrode`
 
@@ -33,19 +213,6 @@
 
     -   update tools script [commit](http://github.com/electrode-io/electrode/commit/5471ef382546c50c6c24d1ac7d33300ee78f78ba)
     -   fix tools update changelog [commit](http://github.com/electrode-io/electrode/commit/4011529aeb22078f8d84704bd39dad57d4d82a2f)
-
-# 6/21/2017
-
-## Packages
-
--   electrode-archetype-react-app-dev@3.0.2 `(3.0.1 => 3.0.2)`
--   electrode-archetype-react-app@3.0.2 `(3.0.1 => 3.0.2)`
-
-## Commits
-
--   `packages/electrode-archetype-react-app[-dev]`
-
-    -   [patch] update peer dep ([#415](https://github.com/electrode-io/electrode/pull/415)) [commit](http://github.com/electrode-io/electrode/commit/3799989a37482c6b094c17f3e2f1e17e283a2b72)
 
 # 6/20/2017